Add files via upload
Marketing Campaigns

Problem scenario:

Marketing mix stands as a widely utilized concept in the execution of marketing 
strategies. It encompasses various facets within a comprehensive marketing plan, 
with a central focus on the four Ps of marketing: product, price, place, and 
promotion.

Problem objective:

As a data scientist, you must conduct exploratory data analysis and hypothesis 
testing to enhance your comprehension of the diverse factors influencing customer 
acquisition.

Data description:

The variables such as birth year, education, income, and others pertain to the first 
'P' or 'People' in the tabular data presented to the user. The expenditures on items 
like wine, fruits, and gold, are associated with ‘Product’. Information relevant to 
sales channels, such as websites and stores, is connected to ‘Place’, and the fields 
discussing promotions and the outcomes of various campaigns are linked to 
‘Promotion’.